requirements.txt
psutil>=5.9.0
mysql-connector-python>=8.0.33
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0